+++ et-common-libs/include/etsoc/isa/cacheops.h
@@ -32,14 +32,13 @@ enum l1d_mode { l1d_shared, l1d_split, l1d_scp };
 //   destination. Optionally, a repeat count can be specified to evict more adjacent cache lines.
 //   Optionally, each potential line eviction can be gated by the value of the TensorMask CSR.
 //
-inline void __attribute__((always_inline))
-evict_sw(uint64_t use_tmask, uint64_t dst, uint64_t way, uint64_t set, uint64_t num_lines)
-{
-    uint64_t csr_enc = ((use_tmask & 1) << 63) | ((dst & 0x3) << 58) | ((set & 0xF) << 14) |
-                       ((way & 0x3) << 6) | (num_lines & 0xF);
-
-    __asm__ __volatile__("csrw 0x7f9, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc));
-}
+#define evict_sw(use_tmask, dst, way, set, num_lines)                                              \
+    {                                                                                              \
+        uint64_t csr_enc = ((((uint64_t)use_tmask) & 1) << 63) | ((((uint64_t)dst) & 0x3) << 58) | \
+                           ((((uint64_t)set) & 0xF) << 14) | ((((uint64_t)way) & 0x3) << 6) |      \
+                           (((uint64_t)num_lines) & 0xF);                                          \
+        __asm__ __volatile__("csrw 0x7f9, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc));               \
+    }
 
 //-------------------------------------------------------------------------------------------------
 //
@@ -49,14 +48,13 @@ evict_sw(uint64_t use_tmask, uint64_t dst, uint64_t way, uint64_t set, uint64_t 
 //   line is dirty. Optionally, a repeat count can be specified to flush more adjacent cache lines.
 //   Optionally, each potential line flush can be gated by the value of the TensorMask CSR.
 //
-inline void __attribute__((always_inline))
-flush_sw(uint64_t use_tmask, uint64_t dst, uint64_t way, uint64_t set, uint64_t num_lines)
-{
-    uint64_t csr_enc = ((use_tmask & 1) << 63) | ((dst & 0x3) << 58) | ((set & 0xF) << 14) |
-                       ((way & 0x3) << 6) | (num_lines & 0xF);
-
-    __asm__ __volatile__("csrw 0x7fb, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc));
-}
+#define flush_sw(use_tmask, dst, way, set, num_lines)                                              \
+    {                                                                                              \
+        uint64_t csr_enc = ((((uint64_t)use_tmask) & 1) << 63) | ((((uint64_t)dst) & 0x3) << 58) | \
+                           ((((uint64_t)set) & 0xF) << 14) | ((((uint64_t)way) & 0x3) << 6) |      \
+                           (((uint64_t)num_lines) & 0xF);                                          \
+        __asm__ __volatile__("csrw 0x7fb, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc));               \
+    }
 
 //-------------------------------------------------------------------------------------------------
 //
@@ -66,12 +64,12 @@ flush_sw(uint64_t use_tmask, uint64_t dst, uint64_t way, uint64_t set, uint64_t 
 //   already locked to another line, or 3 ways are already locked for that set, or the provided
 //   physical address is already present in a different way.
 //
-inline void __attribute__((always_inline)) lock_sw(uint64_t way, uint64_t paddr)
-{
-    uint64_t csr_enc = ((way & 0x3) << 55) | (paddr & 0xFFFFFFFFC0ULL);
-
-    __asm__ __volatile__("csrw 0x7fd, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc));
-}
+#define lock_sw(way, paddr)                                                          \
+    {                                                                                \
+        uint64_t csr_enc = ((((uint64_t)way) & 0x3) << 55) |                         \
+                           (((uint64_t)paddr) & 0xFFFFFFFFC0ULL);                    \
+        __asm__ __volatile__("csrw 0x7fd, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc)); \
+    }
 
 //-------------------------------------------------------------------------------------------------
 //
@@ -79,12 +77,11 @@ inline void __attribute__((always_inline)) lock_sw(uint64_t way, uint64_t paddr)
 //
 //   This function unlocks the provided set and way
 //
-inline void __attribute__((always_inline)) unlock_sw(uint64_t way, uint64_t set)
-{
-    uint64_t csr_enc = ((way & 0xFF) << 55) | ((set & 0xF) << 6);
-
-    __asm__ __volatile__("csrw 0x7ff, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc));
-}
+#define unlock_sw(way, set)                                                                   \
+    {                                                                                         \
+        uint64_t csr_enc = ((((uint64_t)way) & 0xFF) << 55) | ((((uint64_t)set) & 0xF) << 6); \
+        __asm__ __volatile__("csrw 0x7ff, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc));          \
+    }
 
 //-------------------------------------------------------------------------------------------------
 //
@@ -95,18 +92,18 @@ inline void __attribute__((always_inline)) unlock_sw(uint64_t way, uint64_t set)
 //   are calculated using the provided stride.
 //   Optionally, each potential line eviction can be gated by the value of the TensorMask CSR.
 //
-inline void __attribute__((always_inline)) evict_va(uint64_t use_tmask, uint64_t dst, uint64_t addr,
-    uint64_t num_lines, uint64_t stride, uint64_t id)
-{
-    uint64_t csr_enc = ((use_tmask & 1) << 63) | ((dst & 0x3) << 58) | //00=L1, 01=L2, 10=L3, 11=MEM
-                       (addr & 0xFFFFFFFFFFC0ULL) | (num_lines & 0xF);
-
-    register uint64_t x31_enc asm("x31") = (stride & 0xFFFFFFFFFFC0ULL) | (id & 0x1);
-
-    __asm__ __volatile__("csrw 0x89f, %[csr_enc]\n"
-                         :
-                         : [x31_enc] "r"(x31_enc), [csr_enc] "r"(csr_enc));
-}
+#define evict_va(use_tmask, dst, addr, num_lines, stride, id)                                      \
+    {                                                                                              \
+        /* 00=L1, 01=L2, 10=L3, 11=MEM */                                                          \
+        uint64_t csr_enc = ((((uint64_t)use_tmask) & 1ULL) << 63) |                                \
+                           ((((uint64_t)dst) & 0x3ULL) << 58) |                                    \
+                           (((uint64_t)addr) & 0xFFFFFFFFFFC0ULL) | (((uint64_t)num_lines) & 0xF); \
+        register uint64_t x31_enc asm("x31") = (((uint64_t)stride) & 0xFFFFFFFFFFC0ULL) |          \
+                                               (((uint64_t)id) & 0x1ULL);                          \
+        __asm__ __volatile__("csrw 0x89f, %[csr_enc]\n"                                            \
+                             :                                                                     \
+                             : [x31_enc] "r"(x31_enc), [csr_enc] "r"(csr_enc));                    \
+    }
 
 //-------------------------------------------------------------------------------------------------
 //
@@ -115,11 +112,10 @@ inline void __attribute__((always_inline)) evict_va(uint64_t use_tmask, uint64_t
 //   This function evicts all cache lines from address to address+size up to the provided
 //   cache level.
 //
-inline void __attribute__((always_inline))
-evict(enum cop_dest dest, volatile const void *const address, uint64_t size)
-{
-    evict_va(0, dest, (uint64_t)address, (((uint64_t)address & 0x3F) + size) >> 6, 64, 0);
-}
+#define evict(dest, address, size)                                                               \
+    {                                                                                            \
+        evict_va(0, dest, (uint64_t)address, ((((uint64_t)address & 0x3F) + size) >> 6), 64, 0); \
+    }
 
 //-------------------------------------------------------------------------------------------------
 //
@@ -130,20 +126,16 @@ evict(enum cop_dest dest, volatile const void *const address, uint64_t size)
 //   more lines, whose addresses are calculated using the provided stride.
 //   Optionally, each potential line flush can be gated by the value of the TensorMask CSR.
 //
-inline void __attribute__((always_inline)) flush_va(uint64_t use_tmask, uint64_t dst, uint64_t addr,
-    uint64_t num_lines, uint64_t stride, uint64_t id)
-{
-    uint64_t csr_enc = ((use_tmask & 1) << 63) | ((dst & 0x3) << 58) | (addr & 0xFFFFFFFFFFC0ULL) |
-                       (num_lines & 0xF);
-
-    register uint64_t x31_enc asm("x31") = (stride & 0xFFFFFFFFFFC0ULL) | (id & 0x1);
-
-    __asm__ __volatile__("csrw 0x8bf, %[csr_enc]\n"
-                         :
-                         : [x31_enc] "r"(x31_enc), [csr_enc] "r"(csr_enc)
-
-    );
-}
+#define flush_va(use_tmask, dst, addr, num_lines, stride, id)                                      \
+    {                                                                                              \
+        uint64_t csr_enc = ((((uint64_t)use_tmask) & 1) << 63) | ((((uint64_t)dst) & 0x3) << 58) | \
+                           (((uint64_t)addr) & 0xFFFFFFFFFFC0ULL) | (((uint64_t)num_lines) & 0xF); \
+        register uint64_t x31_enc asm("x31") = (((uint64_t)stride) & 0xFFFFFFFFFFC0ULL) |          \
+                                               (((uint64_t)id) & 0x1);                             \
+        __asm__ __volatile__("csrw 0x8bf, %[csr_enc]\n"                                            \
+                             :                                                                     \
+                             : [x31_enc] "r"(x31_enc), [csr_enc] "r"(csr_enc));                    \
+    }
 
 //-------------------------------------------------------------------------------------------------
 //
@@ -154,18 +146,16 @@ inline void __attribute__((always_inline)) flush_va(uint64_t use_tmask, uint64_t
 //   calculated using the provided stride.
 //   Optionally, each line prefetch can be gated by the value of the TensorMask CSR.
 //
-inline void __attribute__((always_inline)) prefetch_va(uint64_t use_tmask, uint64_t dst,
-    uint64_t addr, uint64_t num_lines, uint64_t stride, uint64_t id)
-{
-    uint64_t csr_enc = ((use_tmask & 1) << 63) | ((dst & 0x3) << 58) | (addr & 0xFFFFFFFFFFC0ULL) |
-                       (num_lines & 0xF);
-
-    register uint64_t x31_enc asm("x31") = (stride & 0xFFFFFFFFFFC0ULL) | (id & 0x1);
-
-    __asm__ __volatile__("csrw 0x81f, %[csr_enc]\n"
-                         :
-                         : [x31_enc] "r"(x31_enc), [csr_enc] "r"(csr_enc));
-}
+#define prefetch_va(use_tmask, dst, addr, num_lines, stride, id)                                   \
+    {                                                                                              \
+        uint64_t csr_enc = ((((uint64_t)use_tmask) & 1) << 63) | ((((uint64_t)dst) & 0x3) << 58) | \
+                           (((uint64_t)addr) & 0xFFFFFFFFFFC0ULL) | (((uint64_t)num_lines) & 0xF); \
+        register uint64_t x31_enc asm("x31") = (((uint64_t)stride) & 0xFFFFFFFFFFC0ULL) |          \
+                                               (((uint64_t)id) & 0x1);                             \
+        __asm__ __volatile__("csrw 0x81f, %[csr_enc]\n"                                            \
+                             :                                                                     \
+                             : [x31_enc] "r"(x31_enc), [csr_enc] "r"(csr_enc));                    \
+    }
 
 //-------------------------------------------------------------------------------------------------
 //
@@ -177,17 +167,16 @@ inline void __attribute__((always_inline)) prefetch_va(uint64_t use_tmask, uint6
 //   calculated using the provided stride.
 //   Optionally, each line lock can be gated by the value of the TensorMask CSR.
 //
-inline void __attribute__((always_inline))
-lock_va(uint64_t use_tmask, uint64_t addr, uint64_t num_lines, uint64_t stride, uint64_t id)
-{
-    uint64_t csr_enc = ((use_tmask & 1) << 63) | (addr & 0xFFFFFFFFFFC0ULL) | (num_lines & 0xF);
-
-    register uint64_t x31_enc asm("x31") = (stride & 0xFFFFFFFFFFC0ULL) | (id & 0x1);
-
-    __asm__ __volatile__("csrw 0x8df, %[csr_enc]\n"
-                         :
-                         : [x31_enc] "r"(x31_enc), [csr_enc] "r"(csr_enc));
-}
+#define lock_va(use_tmask, addr, num_lines, stride, id)                                            \
+    {                                                                                              \
+        uint64_t csr_enc = ((((uint64_t)use_tmask) & 1) << 63) |                                   \
+                           (((uint64_t)addr) & 0xFFFFFFFFFFC0ULL) | (((uint64_t)num_lines) & 0xF); \
+        register uint64_t x31_enc asm("x31") = (((uint64_t)stride) & 0xFFFFFFFFFFC0ULL) |          \
+                                               (((uint64_t)id) & 0x1);                             \
+        __asm__ __volatile__("csrw 0x8df, %[csr_enc]\n"                                            \
+                             :                                                                     \
+                             : [x31_enc] "r"(x31_enc), [csr_enc] "r"(csr_enc));                    \
+    }
 
 //-------------------------------------------------------------------------------------------------
 //
@@ -198,17 +187,16 @@ lock_va(uint64_t use_tmask, uint64_t addr, uint64_t num_lines, uint64_t stride, 
 //   calculated using the provided stride.
 //   Optionally, each unlock can be gated by the value of the TensorMask CSR.
 //
-inline void __attribute__((always_inline))
-unlock_va(uint64_t use_tmask, uint64_t addr, uint64_t num_lines, uint64_t stride, uint64_t id)
-{
-    uint64_t csr_enc = ((use_tmask & 1) << 63) | (addr & 0xFFFFFFFFFFC0ULL) | (num_lines & 0xF);
-
-    register uint64_t x31_enc asm("x31") = (stride & 0xFFFFFFFFFFC0ULL) | (id & 0x1);
-
-    __asm__ __volatile__("csrw 0x8ff, %[csr_enc]\n"
-                         :
-                         : [x31_enc] "r"(x31_enc), [csr_enc] "r"(csr_enc));
-}
+#define unlock_va(use_tmask, addr, num_lines, stride, id)                                          \
+    {                                                                                              \
+        uint64_t csr_enc = ((((uint64_t)use_tmask) & 1) << 63) |                                   \
+                           (((uint64_t)addr) & 0xFFFFFFFFFFC0ULL) | (((uint64_t)num_lines) & 0xF); \
+        register uint64_t x31_enc asm("x31") = (((uint64_t)stride) & 0xFFFFFFFFFFC0ULL) |          \
+                                               (((uint64_t)id) & 0x1);                             \
+        __asm__ __volatile__("csrw 0x8ff, %[csr_enc]\n"                                            \
+                             :                                                                     \
+                             : [x31_enc] "r"(x31_enc), [csr_enc] "r"(csr_enc));                    \
+    }
 
 //-------------------------------------------------------------------------------------------------
 //
@@ -216,102 +204,94 @@ unlock_va(uint64_t use_tmask, uint64_t addr, uint64_t num_lines, uint64_t stride
 //
 //   This function invalidates various cache structures used by the minion core.
 //
-inline void __attribute__((always_inline))
-cache_invalidate(uint64_t inval_instr_cache, uint64_t inval_TLBs_and_PTW)
-{
-    uint64_t csr_enc = (inval_TLBs_and_PTW & 1) | ((inval_instr_cache & 1) << 1);
-
-    __asm__ __volatile__("csrw 0x7d0, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc) :);
-}
+#define cache_invalidate(inval_instr_cache, inval_TLBs_and_PTW)                        \
+    {                                                                                  \
+        uint64_t csr_enc = (((uint64_t)inval_TLBs_and_PTW) & 1) |                      \
+                           ((((uint64_t)inval_instr_cache) & 1) << 1);                 \
+        __asm__ __volatile__("csrw 0x7d0, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc) :); \
+    }
 
 //
 // READS CACHE INVALIDATE
 //
-inline uint64_t __attribute__((always_inline)) get_cache_invalidate(void)
-{
-    uint64_t csr_enc;
-    __asm__ __volatile__("csrr %[csr_enc], 0x7d0\n" : [csr_enc] "=r"(csr_enc) : :);
-    return csr_enc;
-}
+#define get_cache_invalidate                                                            \
+    {                                                                                   \
+        uint64_t csr_enc;                                                               \
+        __asm__ __volatile__("csrr %[csr_enc], 0x7d0\n" : [csr_enc] "=r"(csr_enc) : :); \
+        return csr_enc;                                                                 \
+    }
 
 //
 // MCACHE_CONTROL
 //
-inline void __attribute__((always_inline))
-mcache_control(uint64_t d1_split, uint64_t scp_en, uint64_t cacheop_rate, uint64_t cacheop_max)
-{
-    uint64_t csr_enc = ((cacheop_max & 0x1F) << 6) | ((cacheop_rate & 0x7) << 2) |
-                       ((scp_en & 0x1) << 1) | ((d1_split & 0x1) << 0);
-
-    __asm__ __volatile__("csrw 0x7e0, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc) : "x31");
-}
+#define mcache_control(d1_split, scp_en, cacheop_rate, cacheop_max)                             \
+    {                                                                                           \
+        uint64_t csr_enc =                                                                      \
+            ((((uint64_t)cacheop_max) & 0x1F) << 6) | ((((uint64_t)cacheop_rate) & 0x7) << 2) | \
+            ((((uint64_t)scp_en) & 0x1) << 1) | ((((uint64_t)d1_split) & 0x1) << 0);            \
+        __asm__ __volatile__("csrw 0x7e0, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc) : "x31");    \
+    }
 
 //
 // UCACHE_CONTROL
 //
-inline void __attribute__((always_inline))
-ucache_control(uint64_t scp_en, uint64_t cacheop_rate, uint64_t cacheop_max)
-{
-    uint64_t csr_enc = ((cacheop_max & 0x1F) << 6) | ((cacheop_rate & 0x7) << 2) |
-                       ((scp_en & 0x1) << 1);
-
-    __asm__ __volatile__("csrw 0x810, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc) : "x31");
-}
-
-inline enum l1d_mode __attribute__((always_inline)) get_l1d_mode(void)
-{
-    uint64_t csr_enc;
-    __asm__ __volatile__("csrr %[csr_enc], 0x810\n" : [csr_enc] "=r"(csr_enc) : :);
-
-    if ((csr_enc & 0x3) == 0x3)
-    {
-        return l1d_scp;
+#define ucache_control(scp_en, cacheop_rate, cacheop_max)                                    \
+    {                                                                                        \
+        uint64_t csr_enc = ((((uint64_t)cacheop_max) & 0x1F) << 6) |                         \
+                           ((((uint64_t)cacheop_rate) & 0x7) << 2) |                         \
+                           ((((uint64_t)scp_en) & 0x1) << 1);                                \
+        __asm__ __volatile__("csrw 0x810, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc) : "x31"); \
     }
-    else
-    {
-        return ((csr_enc & 0x3) == 0x1) ? l1d_split : l1d_shared;
+
+#define get_l1d_mode                                                                    \
+    {                                                                                   \
+        uint64_t csr_enc;                                                               \
+        __asm__ __volatile__("csrr %[csr_enc], 0x810\n" : [csr_enc] "=r"(csr_enc) : :); \
+        if ((csr_enc & 0x3) == 0x3)                                                     \
+        {                                                                               \
+            return l1d_scp;                                                             \
+        }                                                                               \
+        else                                                                            \
+        {                                                                               \
+            return ((csr_enc & 0x3) == 0x1) ? l1d_split : l1d_shared;                   \
+        }                                                                               \
     }
-}
 
-inline void __attribute__((always_inline)) excl_mode(uint64_t val)
-{
-    __asm__ __volatile__("csrw 0x7d3, %[csr_enc]\n" : : [csr_enc] "r"(val) : "x31");
-}
+#define excl_mode(val)                                                                   \
+    {                                                                                    \
+        __asm__ __volatile__("csrw 0x7d3, %[csr_enc]\n" : : [csr_enc] "r"(val) : "x31"); \
+    }
 
-inline void __attribute__((always_inline)) scp(uint64_t warl, uint64_t DEscratchpad)
-{
-    // Hard partition L1 Data cache between the harts
-    //mcache_control(0,0,0);
-    FENCE
-    WAIT_CACHEOPS
+#define scp(warl, DEscratchpad)                                                              \
+    {                                                                                        \
+        /* Hard partition L1 Data cache between the harts     */                             \
+        /* mcache_control(0,0,0); */                                                         \
+        FENCE                                                                                \
+        WAIT_CACHEOPS                                                                        \
+        /* Enable scratchpad */                                                              \
+        uint64_t csr_enc = ((((uint64_t)warl) & 0x7FFFFFFFFFFFFFFF) << 1) |                  \
+                           (((uint64_t)DEscratchpad) & 0x1);                                 \
+        __asm__ __volatile__("csrw 0x810, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc) : "x31"); \
+    }
 
-    // Enable scratchpad
-    uint64_t csr_enc = ((warl & 0x7FFFFFFFFFFFFFFF) << 1) | (DEscratchpad & 0x1);
-
-    __asm__ __volatile__("csrw 0x810, %[csr_enc]\n" : : [csr_enc] "r"(csr_enc) : "x31");
-}
-
-inline void __attribute__((always_inline)) cb_drain(uint64_t drain_shire, uint64_t drain_bank)
-{
-    // Drain the coalescing buffer of shire cache bank
-    // 1. Write the CB invalidate (assumes FSM always available)
-    volatile uint64_t *sc_idx_cop_sm_ctl_addr =
-        (volatile uint64_t *)ESR_CACHE(drain_shire, drain_bank, SC_IDX_COP_SM_CTL_USER);
-
-    // 2. Checks done
-    uint64_t state;
-    do
-    {
-        state = (*sc_idx_cop_sm_ctl_addr >> 24) & 0xFF;
-    } while (state != 4);
-
-    *sc_idx_cop_sm_ctl_addr = (1 << 0) | // Go bit = 1
-                              (10 << 8); // Opcode = CB_Inv (Coalescing buffer invalidate)
-
-    // 3. Checks done
-    do
-    {
-        state = (*sc_idx_cop_sm_ctl_addr >> 24) & 0xFF;
-    } while (state != 4);
-}
+#define cb_drain(drain_shire, drain_bank)                                                         \
+    {                                                                                             \
+        /* Drain the coalescing buffer of shire cache bank     */                                 \
+        /* 1. Write the CB invalidate (assumes FSM always available) */                           \
+        volatile uint64_t *sc_idx_cop_sm_ctl_addr = (volatile uint64_t *)ESR_CACHE(               \
+            (uint64_t)drain_shire, (uint64_t)drain_bank, SC_IDX_COP_SM_CTL_USER);                 \
+        /* 2. Checks done */                                                                      \
+        uint64_t state;                                                                           \
+        do                                                                                        \
+        {                                                                                         \
+            state = (*sc_idx_cop_sm_ctl_addr >> 24) & 0xFF;                                       \
+        } while (state != 4);                                                                     \
+        *sc_idx_cop_sm_ctl_addr = (1 << 0) | /* Go bit = 1  */                                    \
+                                  (10 << 8); /* Opcode = CB_Inv (Coalescing buffer invalidate) */ \
+        /* 3. Checks done */                                                                      \
+        do                                                                                        \
+        {                                                                                         \
+            state = (*sc_idx_cop_sm_ctl_addr >> 24) & 0xFF;                                       \
+        } while (state != 4);                                                                     \
+    }
 #endif // ! __CACHEOPS_H